Refactor seat selection to seperate method
